+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## dev
+
+
 ## [3.1.0]
+
+
+### Added
+
+- Added support to run the tool without SSL chain verification for users behind proxy servers who act as MITM
 
 ### Changed